First pass at refactorying core/collate.py: Merge extra code in scripts/collate_1d to colalte.py
